Tidy up after running through docs/running-locally.md
