+++ .env.aws.example
@@ -97,7 +97,7 @@ OPENALEX_WORKS_TRANSFORM_MAX_WORKERS=32
 # DMP Works Search
 ######################################
 DMP_WORKS_SEARCH_QUERY_BUILDER=build_dmp_works_search_baseline_query
-DMP_WORKS_SEARCH_RERANK_MODEL=
+# DMP_WORKS_SEARCH_RERANK_MODEL=model_name
 DMP_WORKS_SEARCH_SCROLL_TIME=360m
 DMP_WORKS_SEARCH_BATCH_SIZE=250
 DMP_WORKS_SEARCH_MAX_RESULTS=100
@@ -109,8 +109,8 @@ DMP_WORKS_SEARCH_MAX_CONCURRENT_SHARD_REQUESTS=12
 DMP_WORKS_SEARCH_INNER_HITS_SIZE=50
 DMP_WORKS_SEARCH_ROW_GROUP_SIZE=50000
 DMP_WORKS_SEARCH_ROW_GROUPS_PER_FILE=4
-DMP_WORKS_SEARCH_START_DATE=
-DMP_WORKS_SEARCH_END_DATE=
+# DMP_WORKS_SEARCH_DMPS_START_DATE=YYYY-MM-DD
+# DMP_WORKS_SEARCH_DMPS_END_DATE=YYYY-MM-DD
 
 ######################################
 # OpenSearch Client
