+++ .env.local.example
@@ -33,7 +33,7 @@ UPSTREAM_DATA_CITATION_CORPUS=/path/to/upstream/data_citation_corpus
 # Dataset Subset (Works)
 ########################################################################
 # Enable subset creation to filter works by specific institutions or a list of DOIs.
-DATASET_SUBSET_ENABLE=false
+DATASET_SUBSET_ENABLE=true
 # Path to a JSON file containing a list of ROR IDs and institution names,
 # e.g. [{"name": "University of California, San Diego", "ror": "0168r3w48"}]
 # Works authored by researchers from these institutions will be included.
@@ -46,7 +46,7 @@ DATASET_SUBSET_DOIS_PATH=${DATA_DIR}/meta/work_dois.json
 # Dataset Subset (DMPs)
 ########################################################################
 # Enable subset creation to filter DMPs by specific institutions or a list of DOIs.
-DMP_SUBSET_ENABLE=false
+DMP_SUBSET_ENABLE=true
 # Path to a JSON file containing ROR IDs and institution names.
 # DMPs created by researchers from these institutions will be included.
 DMP_SUBSET_INSTITUTIONS_PATH=${DATA_DIR}/meta/dmp_institutions.json
@@ -91,7 +91,7 @@ OPENALEX_WORKS_TRANSFORM_MAX_WORKERS=32
 # DMP Works Search
 ######################################
 DMP_WORKS_SEARCH_QUERY_BUILDER=build_dmp_works_search_baseline_query
-DMP_WORKS_SEARCH_RERANK_MODEL=
+# DMP_WORKS_SEARCH_RERANK_MODEL=model_name
 DMP_WORKS_SEARCH_SCROLL_TIME=360m
 DMP_WORKS_SEARCH_BATCH_SIZE=250
 DMP_WORKS_SEARCH_MAX_RESULTS=100
@@ -103,8 +103,8 @@ DMP_WORKS_SEARCH_MAX_CONCURRENT_SHARD_REQUESTS=12
 DMP_WORKS_SEARCH_INNER_HITS_SIZE=50
 DMP_WORKS_SEARCH_ROW_GROUP_SIZE=50000
 DMP_WORKS_SEARCH_ROW_GROUPS_PER_FILE=4
-DMP_WORKS_SEARCH_START_DATE=
-DMP_WORKS_SEARCH_END_DATE=
+# DMP_WORKS_SEARCH_DMPS_START_DATE=YYYY-MM-DD
+# DMP_WORKS_SEARCH_DMPS_END_DATE=YYYY-MM-DD
 
 ######################################
 # OpenSearch Client
